requirements.txt
click==7.1.2
fairlearn==0.5.0
graphviz==0.14.2
joblib==0.17.0
jupyter==1.0.0
jupyterlab==2.2.8
matplotlib==3.3.2
multiprocess==0.70.11.1
ohio==0.5.0
pandas==1.1.2
psycopg2==2.8.6
pyyaml==5.3.1
scikit-learn==0.23.2
seaborn==0.11.0
tables==3.6.1
yapf==0.30.0
